extend proposal_serializer
