+++ config/initializers/extends.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'extends/proposal_serializer_extend'
